+++ contracts/modules/GuardianManager.sol
@@ -221,8 +221,7 @@ contract GuardianManager is BaseModule, RelayerModule {
         view
         returns (bool)
     {
-        address signer = recoverSigner(_signHash, _signatures, 0);
-        return isOwner(_wallet, signer); // "GM: signer must be owner"
+        return validateSignatures(_wallet, _signHash, _signatures, OwnerSignature.Required);
     }
 
     function getRequiredSignatures(BaseWallet /* _wallet */, bytes memory _data) internal view returns (uint256) {
